✨ feat: 필요한 event 생성
