--- src/main/java/com/mulmeong/event/ContestWinnerEvent.java
@@ -1,13 +1,0 @@
-package com.mulmeong.event;
-
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.kafka.annotation.KafkaListener;
-
-@Data
-@Getter
-@NoArgsConstructor
-public class ContestWinnerEvent {
-
-}
