--- src/main/java/com/mulmeong/event/FeedCreateEvent.java
@@ -1,12 +1,0 @@
-package com.mulmeong.event;
-
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Data
-@Getter
-@NoArgsConstructor
-public class FeedCreateEvent {
-
-}
