--- src/main/java/com/mulmeong/event/GradeChangeEvent.java
@@ -1,5 +1,0 @@
-package com.mulmeong.event;
-
-public class GradeChangeEvent {
-
-}
