--- src/main/java/com/mulmeong/event/UserChattingCreateEvent.java
@@ -1,4 +1,0 @@
-package com.mulmeong.event;
-
-public class UserChattingCreateEvent {
-}
