+++ src/main/java/com/mulmeong/event/contents/FeedCreatedFollowersEvent.java
@@ -1,0 +1,16 @@
+package com.mulmeong.event.contents;
+
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@Getter
+@NoArgsConstructor
+public class FeedCreatedFollowersEvent {
+    private String feedUuid;
+    private String memberUuid;
+    private List<String> followerUuids;
+}
