+++ src/main/java/com/mulmeong/event/contents/ShortsCommentCreateEvent.java
@@ -1,4 +1,4 @@
-package com.mulmeong.event;
+package com.mulmeong.event.contents;
 
 import lombok.Data;
 import lombok.Getter;
@@ -9,13 +9,10 @@ import java.time.LocalDateTime;
 @Data
 @Getter
 @NoArgsConstructor
-public class ShortsRecommentCreateEvent {
-
+public class ShortsCommentCreateEvent {
+    private String shortsUuid;
     private String memberUuid;
     private String commentUuid;
-    private String recommentUuid;
     private String content;
     private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-
 }
