+++ src/main/java/com/mulmeong/event/contents/ShortsCreatedFollowersEvent.java
@@ -1,0 +1,16 @@
+package com.mulmeong.event.contents;
+
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@Getter
+@NoArgsConstructor
+public class ShortsCreatedFollowersEvent {
+    private String shortsUuid;
+    private String memberUuid;
+    private List<String> followerUuids;
+}
