+++ src/main/java/com/mulmeong/event/member/FollowCreateEvent.java
@@ -1,4 +1,4 @@
-package com.mulmeong.event;
+package com.mulmeong.event.member;
 
 import lombok.Data;
 import lombok.Getter;
@@ -8,5 +8,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class FollowCreateEvent {
-
+    private String sourceUuid;
+    private String targetUuid;
 }
