+++ src/main/java/com/mulmeong/event/member/MemberGradeUpdateEvent.java
@@ -1,0 +1,6 @@
+package com.mulmeong.event.member;
+
+public class MemberGradeUpdateEvent {
+    private String memberUuid;
+    private String grade;
+}
