+++ src/main/java/com/mulmeong/event/report/ReportCreateEvent.java
@@ -1,0 +1,22 @@
+package com.mulmeong.event.report;
+
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Data
+@Getter
+@NoArgsConstructor
+public class ReportCreateEvent {
+    private String memberUuid;
+    private String targetUuid;
+    private ReportType reportType;
+
+    public enum ReportType {
+        MEMBER,
+        FEED,
+        SHORTS,
+        COMMENT,
+        RECOMMENT,
+    }
+}
